requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==1.4.52
python-dotenv==1.0.1
validators==0.22.0
slowapi==0.1.9
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
email-validator==2.1.1
pydantic-settings==2.2.1
git+https://github.com/OpenPecha/Pydurma.git